Update manager_create_role.html
Change the vue for loop selection to vue-multiselect. Spent roughly 1-2hour trying to figure out how to import it correctly and to make it work. Finally found cdn, but later on realize cdn is only for Vue2 after making the multiselect work (The error I get at this point is "selfhook.bind is not a function" and after googling, found out is cause is vue 2 not vue3). Spent more time and  finally found the cdn for vue 3 in an issue post in github instead of being in the documentation.