Attempt to recode parts of code to make sure role id and srbr are auto increment
Right now role can be added with no problem, but I am not able to return the new role id back into the html, even after a whole day of attempts. Might have something to do with the request being post instead of get. A new function to get role by timestamp can be created after the timestamp column is added to the db so a get request can be used to grab the new role id for creation of skill required by role.