Work done to convert current rolewy.html to vue. Read description for more detail.
Code is added to manager_all_role.html to prepare the vue model for the role page.

Currently, the vue model is working fine except for one part: When using a for loop to call axios to fetch all srbr with role id, the axios only use the last value in the loop. This need to be fixed if vue is to be used for the page.

Additional stuff: Fix code for getting all srbr with role id.