Fixed modal only showing info of first role
The error is that I have created a modal for each role, so when the button to call for modal is pressed, all the modals are called and the first role modal is the one all the way at the top covering the rest of them. Now, I have change the id of each modal to include the role id so only one of them is called at any one time.